+++ packages/card/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - ab1fd9e: Updates Card component for Visual brand Refresh
 
+### Patch Changes
+- Updated dependencies [ab1fd9e]
+  - @leafygreen-ui/tokens@1.0.0
+
 ## 5.1.4
 
 ### Patch Changes